refactor: split API code into smaller units
Moved the big `api/validation` function to its own file. This work paves
the work for the `api/audit` function.
